+++ react/src/components/AtomForm/styled/AtomFormCardRow.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export default styled.div`
+  display: flex;
+  position: relative;
+  z-index: 2;
+`